# Fix race condition in BinaryReader.LookupSymbol()