--- a/pkg/block/indexheader/binary_reader.go
+++ b/pkg/block/indexheader/binary_reader.go
@@ -15,6 +15,7 @@ import (
 	"os"
 	"path/filepath"
 	"sort"
+	"sync"
 	"time"
 	"unsafe"
 
@@ -440,7 +441,8 @@ type BinaryReader struct {
 	nameSymbols map[uint32]string
 	// Direct cache of values. This is much faster than an LRU cache and still provides
 	// a reasonable cache hit ratio.
-	valueSymbols [valueSymbolsCacheSize]struct {
+	valueSymbolsMx sync.Mutex
+	valueSymbols   [valueSymbolsCacheSize]struct {
 		index  uint32
 		symbol string
 	}
@@ -812,9 +814,13 @@ func (r BinaryReader) postingsOffset(name string, values ...string) ([]index.Ran
 
 func (r *BinaryReader) LookupSymbol(o uint32) (string, error) {
 	cacheIndex := o % valueSymbolsCacheSize
+	r.valueSymbolsMx.Lock()
 	if cached := r.valueSymbols[cacheIndex]; cached.index == o && cached.symbol != "" {
-		return cached.symbol, nil
+		v := cached.symbol
+		r.valueSymbolsMx.Unlock()
+		return v, nil
 	}
+	r.valueSymbolsMx.Unlock()
 
 	if s, ok := r.nameSymbols[o]; ok {
 		return s, nil
@@ -830,8 +836,12 @@ func (r *BinaryReader) LookupSymbol(o uint32) (string, error) {
 	if err != nil {
 		return s, err
 	}
+
+	r.valueSymbolsMx.Lock()
 	r.valueSymbols[cacheIndex].index = o
 	r.valueSymbols[cacheIndex].symbol = s
+	r.valueSymbolsMx.Unlock()
+
 	return s, nil
 }