* [x] I added CHANGELOG entry for this change.
* [ ] Change is not relevant to the end user.

## Changes

The symbols cache introduced in https://github.com/thanos-io/thanos/pull/3557 also introduced a subtle race condition, because `LookupSymbol()` may be called concurrently.

In this PR I'm proposing to introduce a lock. With a lock it's obviously slower, but benchmarks show it may not be much significative compared to not having a cache at all (it's late, maybe I'm missing something, so be critic on this please).

### PR compared to `master`

```
benchmark                                                    old ns/op     new ns/op     delta
BenchmarkBinaryReader_LookupSymbol/num_series_=_1024-12      3288          13786         +319.28%
BenchmarkBinaryReader_LookupSymbol/num_series_=_10240-12     1850285       1996616       +7.91%

benchmark                                                    old allocs     new allocs     delta
BenchmarkBinaryReader_LookupSymbol/num_series_=_1024-12      0              0              +0.00%
BenchmarkBinaryReader_LookupSymbol/num_series_=_10240-12     10230          10230          +0.00%

benchmark                                                    old bytes     new bytes     delta
BenchmarkBinaryReader_LookupSymbol/num_series_=_1024-12      0             1             +Inf%
BenchmarkBinaryReader_LookupSymbol/num_series_=_10240-12     40576         40608         +0.08%
```

### Removing the symbols cache at all compared to `master`

```
benchmark                                                    old ns/op     new ns/op     delta
BenchmarkBinaryReader_LookupSymbol/num_series_=_1024-12      3288          179712        +5365.69%
BenchmarkBinaryReader_LookupSymbol/num_series_=_10240-12     1850285       1817189       -1.79%

benchmark                                                    old allocs     new allocs     delta
BenchmarkBinaryReader_LookupSymbol/num_series_=_1024-12      0              1014           +Inf%
BenchmarkBinaryReader_LookupSymbol/num_series_=_10240-12     10230          10230          +0.00%

benchmark                                                    old bytes     new bytes     delta
BenchmarkBinaryReader_LookupSymbol/num_series_=_1024-12      0             3179          +Inf%
BenchmarkBinaryReader_LookupSymbol/num_series_=_10240-12     40576         40572         -0.01%
```

## Verification

Unit tests / benchmark.